Instantiate concrete type macros directly in ginger
