+++ type-mappings/src/instantiated/mod.rs
@@ -1,0 +1,12 @@
+#[cfg(all(not(feature = "bn_382"), feature = "tweedle"))]
+pub mod tweedle;
+#[cfg(all(not(feature = "bn_382"), feature = "tweedle"))]
+pub use tweedle::*;
+
+#[cfg(all(not(feature = "tweedle"), feature = "bn_382"))]
+pub mod bn_382;
+#[cfg(all(not(feature = "tweedle"), feature = "bn_382"))]
+pub use bn_382::*;
+
+#[allow(unused_imports)]
+use crate::macros::*;
